requirements.txt
pandas
numpy
matplotlib
openpyxl
sqlalchemy
jinja2
weasyprint
openai
python-dotenv